feat(cli)!: Ditch zerover, rock it with semver
[BREAKING CHANGE]: Nothing burger.

There isn’t any good reason to keep this project in zerover limbo any
more. It isn’t “done” in that there are still features that could be
added; but it isn’t changing with the winds either. The existing
feature set is stable and useful in production and usage is not
expected to change radically. If the API were to remove a method or an
existing CLI flag to change or something we’ll mark it as a breaking
change again. In the mean time, a semver based scheme of minor releases
and patch releases will serve the project better.